Parse Juniper level-specific passive IS-IS setting
