+++ projects/batfish/src/main/antlr4/org/batfish/grammar/flatjuniper/FlatJuniper_isis.g4
@@ -128,6 +128,7 @@ isi_level
     | isil_hello_interval
     | isil_hold_time
     | isil_metric
+    | isil_passive
     | isil_priority
     | isil_te_metric
   )
@@ -197,6 +198,11 @@ isil_hold_time
 isil_metric
 :
   METRIC DEC
+;
+
+isil_passive
+:
+  PASSIVE
 ;
 
 isil_priority
