+++ projects/batfish/src/main/java/org/batfish/representation/juniper/JuniperConfiguration.java
@@ -665,7 +665,9 @@ public final class JuniperConfiguration extends VendorConfiguration {
         .setHelloInterval(interfaceLevelSettings.getHelloInterval())
         .setHoldTime(interfaceLevelSettings.getHoldTime())
         .setMode(
-            interfaceSettings.getPassive() ? IsisInterfaceMode.PASSIVE : IsisInterfaceMode.ACTIVE)
+            interfaceLevelSettings.getPassive()
+                ? IsisInterfaceMode.PASSIVE
+                : IsisInterfaceMode.ACTIVE)
         .build();
   }
